// frontend/src/app/dashboard/certified-copy/general/my-cases/interfaces/all-case-details.ts
export interface AllCaseDetails{
    case_no: string;
    case_type_name: string;
    cino: string;
    datevalue: string;
    filing_no: string;
    finalcaseno: string;
    next_date: string;
    pet_address: string;
    pet_adv_name: string;
    pet_name: string;
    regis_date: string;
    res_address: string;
    res_adv_name: string;
    res_name: string;
    short_name: string; 
    status: string;
}